dont use float

+++ src/handlers.c
@@ -81,8 +81,7 @@ handler_button_press(XEvent *xevent) {
         clock_gettime(CLOCK_MONOTONIC, &tnow);
         if (timediff(tnow, xsel.tclick2) <= CONF_TRIPLE_CLICK_TIMEOUT) {
             snap = SELECTION_SNAP_LINE;
-        } else if (timediff(tnow, xsel.tclick1)
-                   <= (float)CONF_DOUBLE_CLICK_TIMEOUT) {
+        } else if (timediff(tnow, xsel.tclick1) <= CONF_DOUBLE_CLICK_TIMEOUT) {
             snap = SELECTION_SNAP_WORD;
         } else {
             snap = 0;
